Fixed GCC warnings at -Wall level.
This mainly fixes those nasty signed/unsigned comparison warnings.

+++ gf/test/gf_new_test.cpp
@@ -21,7 +21,7 @@ TEST(GreensFunction, InitializationTest){
   alps::gf::legendre_mesh w(100, 10);
 
   greenf<double, alps::gf::matsubara_positive_mesh, alps::gf::index_mesh, alps::gf::itime_mesh, alps::gf::legendre_mesh> g(x, y, z, w);
-  for(int i = 0; i<g.data().size(); ++i) {
+  for(size_t i = 0; i<g.data().size(); ++i) {
     ASSERT_NEAR(g.data().storage().data(i), 0.0, 1E-15);
   }
   for(alps::gf::matsubara_positive_mesh::index_type i(0); i<x.extent(); ++i) {
